Move CourseEditor into CourseForm component
